+++ drift/lib/src/runtime/query_builder/components/order_by.dart
@@ -37,6 +37,12 @@ class OrderingTerm extends Component {
     return OrderingTerm(expression: expression, mode: OrderingMode.desc);
   }
 
+  /// Creates an ordering term  to get a number of random rows
+  /// using sqlite random function.
+  factory OrderingTerm.random() {
+    return OrderingTerm(expression: FunctionCallExpression('random', []));
+  }
+
   @override
   void writeInto(GenerationContext context) {
     expression.writeInto(context);